+++ src/components/SalesProductTitle.js
@@ -1,4 +1,4 @@
-import SortOrder from "./SortOredr";
+import SortOrder from "./SortOrder";
 import '../css/SalesProductTitle.css';
 import { useState } from "react";
 
@@ -18,16 +18,18 @@ export default function SalesProductTitle({ onSearchChange, onSortOrderChange })
   return (
     <div className="salesProductTitle">
       <h1 className='title'>판매 중인 상품</h1>
-      <input
-      className="searchArea"
-      type="text"
-      placeholder="검색할 상품을 입력해주세요"
-      value={searchText}
-      onChange={handleSearchChange}
-      onKeyDown={handleKeyDown}
-      ></input>
-      <button className="registButton">상품 등록하기</button>
-      <SortOrder onChange={onSortOrderChange} />
+      <div className="titleFn">
+        <input
+        className="searchArea"
+        type="text"
+        placeholder="검색할 상품을 입력해주세요"
+        value={searchText}
+        onChange={handleSearchChange}
+        onKeyDown={handleKeyDown}
+        ></input>
+        <button className="registButton">상품 등록하기</button>
+        <SortOrder onChange={onSortOrderChange} />
+      </div>
     </div>
   )
 }
