+++ src/css/App.css
@@ -1,13 +1,13 @@
 html {
   /* 10px */
   font-size: 62.5%;
-  font-family: Pretendard;
 }
 
 * {
   box-sizing: border-box;
   margin: 0;
   padding: 0;
+  font-family: Pretendard;
 }
