+++ src/css/Nav.css
@@ -31,3 +31,16 @@ header {
   color: #4B5563;
   padding: 2.4rem 1.6rem
 }
+
+.user {
+  display: flex;
+  align-items: center;
+  gap: 0.6rem;
+}
+
+.user p {
+  color: #4B5563;
+  font-size: 1.8rem;
+  font-weight: 700;
+  line-height: 2.6rem;
+}
