+++ src/css/Pagination.css
@@ -1,0 +1,34 @@
+.paginationWrapper {
+  display: flex;
+  justify-content: center;
+  gap: 0.4rem;
+  margin: 4.3rem 0 14rem 0;
+}
+
+button {
+  width: 4rem;
+  height: 4rem;
+  font-weight: 600;
+  background-color: #ffffff;
+  border-radius: 99rem;
+  border: 0.1rem solid #E5E7EB;
+  font-size: 1.6rem;
+  line-height: 2.6rem;
+  color: #6B7280;
+}
+
+.pageBtns {
+  display: flex;
+  gap: 0.4rem;
+}
+
+.pageBtns button.active {
+  background-color: #2F80ED;
+  color: #F9FAFB;
+  border: none;
+}
+
+.pageBtns button:disabled {
+  cursor: not-allowed;
+  background-color: #ffffff;
+}
