+++ src/css/ProductItem.css
@@ -1,7 +1,32 @@
-.img {
+.BestProduct  {
+  display: flex;
+  flex-direction: column;
+  gap: 1.6rem;
+}
+
+.BestProduct .img {
   border-radius: 1.6rem;
   width: 28.2rem;
   height: 28.2rem;
+}
+
+.SalesProduct {
+  display: flex;
+  flex-direction: column;
+  gap: 1.6rem;
+}
+
+.SalesProduct .img {
+  border-radius: 1.6rem;
+  width: 22.1rem;
+  height: 22.1rem;
+}
+
+.textSection {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  color: #1F2937;
 }
 
 .name {
@@ -15,3 +40,11 @@ font-size: 1.6rem;
 font-weight: 700;
 line-height: 2.6rem;
 }
+
+.favoriteCount {
+  font-size: 1.2rem;
+  font-weight: 500;
+  line-height: 1.8rem;
+  color: #4B5563;
+}
+
