+++ src/css/SalesProduct.css
@@ -1,30 +1,8 @@
-.title {
-  font-size: 2rem;
-  font-weight: 700;
-  line-height: 3.2rem;
-}
-
-.img {
-  border-radius: 1.6rem;
-  width: 28.2rem;
-  height: 28.2rem;
-}
-
-.name {
-  font-size: 1.4rem;
-  font-weight: 500;
-  line-height: 2.4rem;
-}
-
-.price {
-  font-size: 1.6rem;
-  font-weight: 700;
-  line-height: 2.6rem;
-}
-
 .salesItems {
-  display: flex;
+  display: grid;
+  grid-template-columns: repeat(5, 1fr);
   gap: 2.4rem;
+  max-width: 120rem;
   list-style-type: none;
 }
