+++ src/css/SalesProductTitle.css
@@ -1,5 +1,11 @@
 .salesProductTitle {
   display: flex;
+  justify-content: space-between;
+}
+
+.titleFn {
+  display: flex;
+  gap: 1.2rem;
 }
 
 .searchArea {
@@ -12,14 +18,18 @@
 }
 
 .registButton {
+  display: flex;
+  align-items: center;
+  justify-content: center;
   width: 13.3rem;
   height: 4.2rem;
-  padding: 1.2rem 2.3rem;
   border: none;
   border-radius: 0.8rem;
   background-color: #3692FF;
+  font-family: pretendard;
   font-size: 1.6rem;
   font-weight: 600;
   line-height: 2.6rem;
   color: #F3F4F6;
+  cursor: pointer;
 }
