+++ src/css/SortOrder.css
@@ -1,0 +1,28 @@
+.sortOder {
+  display: flex;
+  align-items: center;
+  position: relative;
+}
+
+select {
+  -moz-appearance: none;
+  -webkit-appearance: none;
+  appearance: none;
+  width: 13rem;
+  height: 4.2rem;
+  border: 0.1rem solid #E5E7EB;
+  border-radius: 1.2rem;
+  padding: 0 2rem 0 2rem;
+  font-family: Pretendard;
+  font-size: 1.6rem;
+  font-weight: 400;
+}
+
+.optionText {
+  font-weight: 400;
+}
+
+.dropdownArrowImg {
+  position: absolute;
+  right: 2rem;
+}
